requirements.txt
# ASMA-SLED: Advanced Sound Event Localization and Detection
# Production-ready requirements for professional deployment

# Core Dependencies
torch>=1.13.0
torchaudio>=0.13.0
torchvision>=0.14.0
numpy>=1.21.0
scipy>=1.8.0

# Audio Processing
librosa>=0.9.0
soundfile>=0.10.3
audioread>=3.0.0

# Video Processing  
opencv-python>=4.6.0
pillow>=9.0.0
imageio>=2.19.0
imageio-ffmpeg>=0.4.7

# Scientific Computing
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# Configuration Management
pyyaml>=6.0
hydra-core>=1.2.0
omegaconf>=2.2.0

# Utilities
tqdm>=4.64.0
tensorboard>=2.9.0
h5py>=3.7.0
joblib>=1.1.0

# Development Tools
pytest>=7.1.0
black>=22.3.0
flake8>=4.0.0
isort>=5.10.0

# Optional GPU Acceleration (install manually with CUDA version)
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# System Requirements:
# - Python 3.8+
# - CUDA 11.0+ (for GPU training)
# - 16GB+ RAM recommended
# - 50GB+ storage for dataset